+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `/forums`, `/obb`, `/gallery`, and `/exchange` were simplified into `/link`.
 * `/info` no longer checks for the user's staff role, and instead only checks the category name of the channel the command was called in.
 * `/wiki` was split into separate commands: `/wikilink`, `/wikisearch`, and `/wikisummary`.
+* Users who are repeatedly spamming/using caps are now warned through direct messages, rather than in public channels.
 
 ### Removed
 * The exalt/unexalt commands were removed.